refactor: phase 01 polish and hardening
